# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for accessing orienteering competition data from MeOS",
        "title": "meos-graphics",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "@malpou",
            "url": "https://github.com/MetsaApp/meos-graphics",
            "email": "malthe@grundtvigsvej.dk"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.2.0 x-release-please-version"
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
        "/classes": {
            "get": {
                "description": "Get a list of all competition classes sorted by order key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "Get all competition classes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.ClassInfo"
                            }
                        }
                    }
                }
            }
        },
        "/classes/{classId}/results": {
            "get": {
                "description": "Get the results for a specific competition class including positions and times",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "Get results for a class",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Class ID",
                        "name": "classId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.ResultEntry"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/classes/{classId}/splits": {
            "get": {
                "description": "Get split times at each control for a specific competition class",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "Get split times for a class",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Class ID",
                        "name": "classId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.SplitsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/classes/{classId}/startlist": {
            "get": {
                "description": "Get the start list for a specific competition class",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "Get start list for a class",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Class ID",
                        "name": "classId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.StartListEntry"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.ClassInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "orderKey": {
                    "type": "integer"
                }
            }
        },
        "service.ResultEntry": {
            "type": "object",
            "properties": {
                "club": {
                    "type": "string"
                },
                "difference": {
                    "description": "Formatted duration from leader",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "runningTime": {
                    "description": "Formatted duration string",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "service.SplitStanding": {
            "type": "object",
            "properties": {
                "controlId": {
                    "type": "integer"
                },
                "controlName": {
                    "type": "string"
                },
                "standings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.SplitTime"
                    }
                }
            }
        },
        "service.SplitTime": {
            "type": "object",
            "properties": {
                "club": {
                    "type": "string"
                },
                "elapsedTime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "timeDifference": {
                    "type": "string"
                }
            }
        },
        "service.SplitsResponse": {
            "type": "object",
            "properties": {
                "className": {
                    "type": "string"
                },
                "splits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.SplitStanding"
                    }
                }
            }
        },
        "service.StartListEntry": {
            "type": "object",
            "properties": {
                "club": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startTime": {
                    "description": "Formatted as HH:mm",
                    "type": "string"
                }
            }
        }
    }
}